Remove redundant declaration of util_alloc_realpath()
Function util_alloc_realpath() is, following commit c4b0b8a, declared in
two locations--lines 115 and 299.  The declaration on line 115 was
originally introduced as part of commit 460b240, while the declaration
on line 299 was introduced as a side effect, possibly inadvertently,
of defining the util_exit() function in commit 7632fba4.

This commit removes the declaration on line 299, but this can be easily
restored later if that location is more appropriate.