ERT-747: Changed fault_block_list_neighbours()
The original implementation of fault_block_list_neighbours() detected
the edge of block, and checked all cells on the immediate outside of the
block. This new implementation fully ignores the edge concept and just
scans through all the cells in the block, and checks the immediate
cartesian neighbours.

The reason for the change is that the old implementation had problems
with blocks which were fully embedded in other blocks.